+++ app/component/MainMenu.js
@@ -11,6 +11,7 @@ import MainMenuLinks from './MainMenuLinks';
 import { addAnalyticsEvent } from '../util/analyticsUtils';
 import { updateCountries } from '../action/CountryActions';
 import Toggle from './customizesearch/Toggle';
+import { configureCountry } from '../util/configureCountry';
 
 function MainMenu(props, { config, intl }) {
   const [countries, setCountries] = useState(props.countries);
@@ -101,6 +102,10 @@ function MainMenu(props, { config, intl }) {
                         [country]: !countries[country],
                       });
                       props.updateCountries({
+                        ...countries,
+                        [country]: !countries[country],
+                      });
+                      configureCountry(config, {
                         ...countries,
                         [country]: !countries[country],
                       });
